CASMPET-4199: Run unit tests in container with go
The unit tests failed on a build. DST said that it failed due to
not having go on the build node. They suggested that the unit
tests should be run in a docker container rather than directly on
the node via using docker run. So that's what's implemented here.

I verified this worked by setting WORKSPACE=$(pwd), and running
`./runUnitTest.sh`. I forced a unit test failure and it exited
with non-zero.